requirements.txt
flask
gunicorn
flask_sqlalchemy
psycopg2-binary
flask-migrate
python-dotenv
PyJWT
werkzeug
boto3
redis
rq
Flask-JWT-Extended==4.6.0
requests
pytest
pytest-cov
flake8
flask-cors
Flask-Limiter[redis]
Flask-Mail
email-validator
marshmallow
flask-marshmallow
cryptography
